requirements.txt
alembic==1.9.4
black==23.1.0
cffi==1.15.1
click==8.1.3
cryptography==39.0.2
Deprecated==1.2.13
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
python-dotenv==1.0.0
python-jwt==4.0.0
six==1.16.0
SQLAlchemy==2.0.4
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
wrapt==1.15.0
